Added some services, attempting hunter and gather, not working yet for deploy.
