+++ Utilities/devutils/constantclipboard.sh
@@ -2,7 +2,7 @@
 
 # Use this as a part of a daemon to keep the clipboard active.
 
-while true; daemon
+while true; do
     sudo VBoxClient --clipboard;
     sudo VBoxClient --draganddrop;
     sleep 60;
