+++ Utilities/systemd-services/constantclipboard.service
@@ -1,0 +1,12 @@
+[Unit]
+Description=Enable VirtualBox Clipboard and Drag and Drop features
+
+[Service]
+User=YOURUSERNAMEORWHATEVER
+WorkingDirectory=/path/to/where/it/is (not the full path to the file directly)
+Type=simple
+ExecStart=/path/to/the/file/like/constantclipboard.sh
+Restart=always
+
+[Install]
+WantedBy=multi-user.target
