+++ WebServer/_webserver.log
@@ -21,3 +21,7 @@
 {"type":"log","source":"_SPOTIFY_","date":"2023-11-16T00:44:32.282EST","log_message":"User requested some jams, got some.\n"}
 {"type":"log","source":"Webserver","date":"2023-11-17T17:37:08.066EST","log_message":"User logged in successfully.\n"}
 {"type":"log","source":"_SPOTIFY_","date":"2023-11-17T17:37:28.394EST","log_message":"User requested some jams, got some.\n"}
+{"type":"log","source":"Webserver","date":"2023-11-20T11:30:11.494EST","log_message":"User logged in successfully.\n"}
+{"type":"log","source":"_SPOTIFY_","date":"2023-11-20T11:33:46.848EST","log_message":"User requested some jams, got some.\n"}
+{"type":"log","source":"_REGISTRATION_","date":"2023-11-20T11:36:30.477EST","log_message":"[REGISTRATION] User registered successfully.\n"}
+{"type":"log","source":"Webserver","date":"2023-11-20T11:37:47.462EST","log_message":"User logged in successfully.\n"}
